refactor to expose builders for individual package layers
