+++ ftl/php/builder.py
@@ -13,16 +13,13 @@
 # limitations under the License.
 """This package defines the interface for orchestrating image builds."""
 
-import os
-import subprocess
-import tempfile
-import datetime
+import logging
 import json
 
 from ftl.common import builder
 from ftl.common import ftl_util
-from ftl.common import single_layer_image
-from ftl.common import tar_to_dockerimage
+from ftl.common import layer_builder as base_builder
+from ftl.php import layer_builder as php_builder
 
 _PHP_NAMESPACE = 'php-package-lock-cache'
 _COMPOSER_LOCK = 'composer.lock'
@@ -52,25 +49,34 @@ class PHP(builder.RuntimeBase):
             if len(pkgs) > 41:
                 pkgs = [None]
             for pkg_txt in pkgs:
-                pkg = self.PackageLayer(self._ctx, self._descriptor_files,
-                                        pkg_txt, self._args.destination_path)
+                logging.info('building package layer')
+                logging.info(pkg_txt)
+                layer_builder = php_builder.LayerBuilder(
+                    ctx=self._ctx,
+                    descriptor_files=self._descriptor_files,
+                    pkg_descriptor=pkg_txt,
+                    destination_path=self._args.destination_path)
                 cached_pkg_img = None
                 if self._args.cache:
                     with ftl_util.Timing("checking cached pkg layer"):
-                        cached_pkg_img = self._cache.Get(pkg.GetCacheKey())
+                        key = layer_builder.GetCacheKey()
+                        cached_pkg_img = self._cache.Get(key)
                 if cached_pkg_img is not None:
-                    pkg.SetImage(cached_pkg_img)
+                    layer_builder.SetImage(cached_pkg_img)
                 else:
                     with ftl_util.Timing("building pkg layer"):
-                        pkg.BuildLayer()
+                        layer_builder.BuildLayer()
                     if self._args.cache:
                         with ftl_util.Timing("uploading pkg layer"):
-                            self._cache.Set(pkg.GetCacheKey(), pkg.GetImage())
-                lyr_imgs.append(pkg)
+                            self._cache.Set(layer_builder.GetCacheKey(),
+                                            layer_builder.GetImage())
+                lyr_imgs.append(layer_builder)
 
-        app = self.AppLayer(self._ctx, self._args.destination_path,
-                            self._args.entrypoint,
-                            self._args.exposed_ports)
+        app = base_builder.AppLayerBuilder(
+            ctx=self._ctx,
+            destination_path=self._args.destination_path,
+            entrypoint=self._args.entrypoint,
+            exposed_ports=self._args.exposed_ports)
         with ftl_util.Timing("builder app layer"):
             app.BuildLayer()
         lyr_imgs.append(app)
@@ -78,60 +84,3 @@ class PHP(builder.RuntimeBase):
             ftl_image = self.AppendLayersIntoImage(lyr_imgs)
         with ftl_util.Timing("uploading final image"):
             self.StoreImage(ftl_image)
-
-    class PackageLayer(single_layer_image.CacheableLayer):
-        def __init__(self, ctx, descriptor_files, pkg_descriptor,
-                     destination_path):
-            super(PHP.PackageLayer, self).__init__()
-            self._ctx = ctx
-            self._descriptor_files = descriptor_files
-            self._pkg_descriptor = pkg_descriptor
-            self._destination_path = destination_path
-
-        def GetCacheKeyRaw(self):
-            if self._pkg_descriptor is not None:
-                # phase 2 cache key
-                return self._pkg_descriptor[0] + ' ' + self._pkg_descriptor[1]
-            # phase 1 cache key
-            return ftl_util.descriptor_parser(self._descriptor_files,
-                                              self._ctx)
-
-        def BuildLayer(self):
-            """Override."""
-            blob, u_blob = self._gen_composer_install_tar(
-                self._pkg_descriptor, self._destination_path)
-            overrides_dct = {
-                    'created': str(datetime.date.today()) + "T00:00:00Z"
-                }
-            self._img = tar_to_dockerimage.FromFSImage(
-                [blob], [u_blob], overrides_dct)
-
-        def _gen_composer_install_tar(self, pkg_descriptor, destination_path):
-            # Create temp directory to write package descriptor to
-            pkg_dir = tempfile.mkdtemp()
-            app_dir = os.path.join(pkg_dir, destination_path.strip("/"))
-            os.makedirs(app_dir)
-
-            # Copy out the relevant package descriptors to a tempdir.
-            if pkg_descriptor is None:
-                # phase 1 copy whole descriptor
-                ftl_util.descriptor_copy(self._ctx, self._descriptor_files,
-                                         app_dir)
-
-            subprocess.check_call(
-                ['rm', '-rf', os.path.join(app_dir, 'vendor')])
-
-            with ftl_util.Timing("composer_install"):
-                if pkg_descriptor is None:
-                    # phase 1 install entire descriptor
-                    subprocess.check_call(
-                        ['composer', 'install', '--no-dev', '--no-scripts'],
-                        cwd=app_dir)
-                else:
-                    pkg, version = pkg_descriptor
-                    subprocess.check_call(
-                        ['composer', 'require',
-                         str(pkg),
-                         str(version)],
-                        cwd=app_dir)
-            return ftl_util.zip_dir_to_layer_sha(pkg_dir)
